+++ test_projects/python/source.py
@@ -9,5 +9,9 @@ class Klass:
     def method(self):
         funktion(KONSTANT)
 
+    def method_with_underscores(self):
+        ...
+
 
 Klass().method()
+Klass().method_with_underscores()